--- a/src/Compression/CompressedReadBufferBase.cpp
+++ b/src/Compression/CompressedReadBufferBase.cpp
@@ -232,8 +232,8 @@ void CompressedReadBufferBase::decompress(BufferBase::Buffer & to, size_t size_d
         UInt8 header_size = ICompressionCodec::getHeaderSize();
         if (size_compressed_without_checksum < header_size)
             throw Exception(ErrorCodes::CORRUPTED_DATA,
-                "Can't decompress data: the compressed data size ({}), this should include header size) is less than the header size ({})",
-                    size_compressed_without_checksum, size_t(header_size));
+                "Can't decompress data: the compressed data size ({}, this should include header size) is less than the header size ({})",
+                    size_compressed_without_checksum, static_cast<size_t>(header_size));
 
         to = BufferBase::Buffer(compressed_buffer + header_size, compressed_buffer + size_compressed_without_checksum);
     }