… tests/queries/1_stateful/00162_mmap_compression_none.reference