@alexey-milovidov 2021-03-25
Two times performance improvement:

```
CREATE TABLE test.hits_none (Title String CODEC(NONE)) ENGINE = MergeTree ORDER BY tuple();
INSERT INTO test.hits_none SELECT Title FROM test.hits;
OPTIMIZE TABLE test.hits_none FINAL;
```

```
clickhouse-benchmark --query "SELECT sum(length(Title)) FROM test.hits_none"
```

Before: 
```
localhost:9000, queries 210, QPS: 7.624, RPS: 67652204.332, MiB/s: 5576.078, result RPS: 7.624, result MiB/s: 0.000.

0.000%          0.114 sec.
10.000%         0.122 sec.
20.000%         0.124 sec.
30.000%         0.126 sec.
40.000%         0.128 sec.
50.000%         0.130 sec.
```

After:
```
localhost:9000, queries 248, QPS: 15.142, RPS: 134366240.938, MiB/s: 11074.830, result RPS: 15.142, result MiB/s: 0.000.

0.000%          0.055 sec.
10.000%         0.059 sec.
20.000%         0.061 sec.
30.000%         0.063 sec.
40.000%         0.065 sec.
50.000%         0.066 sec.
```

PS. Compressed hits table is faster on my machine:
```
localhost:9000, queries 597, QPS: 21.958, RPS: 194853523.858, MiB/s: 16060.356, result RPS: 21.958, result MiB/s: 0.000.

0.000%          0.038 sec.
10.000%         0.041 sec.
20.000%         0.042 sec.
30.000%         0.043 sec.
40.000%         0.044 sec.
50.000%         0.045 sec.
```