@alexey-milovidov 2021-03-25
Something is wrong with mmap IO: 
```
clickhouse-benchmark --min_bytes_to_use_mmap_io 1 --max_threads 1 --query "SELECT sum(length(Title)) FROM test.hits_none"
```