@alexey-milovidov 2021-03-26
TODO:
- add perf test;
- add scenario with MMap IO to test.