@alexey-milovidov 2021-03-27
Performance test has shown the improvement.